refactor: move RenderOutput DTO to model/shared, re-export from renderer port
