+++ src/app/ports/outbound/renderer.rs
@@ -2,8 +2,9 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use crate::model::app_state::AppState;
-use crate::model::shared::viewport::{ColumnWidthsCache, ViewportPlan};
 use crate::services::AppServices;
+
+pub use crate::model::shared::render_output::RenderOutput;
 
 #[derive(Debug, Clone, thiserror::Error)]
 pub enum RenderError {
@@ -18,30 +19,6 @@ impl From<std::io::Error> for RenderError {
 }
 
 pub type RenderResult<T> = Result<T, RenderError>;
-
-#[derive(Default)]
-pub struct RenderOutput {
-    pub inspector_viewport_plan: ViewportPlan,
-    pub result_viewport_plan: ViewportPlan,
-    pub result_widths_cache: ColumnWidthsCache,
-    pub explorer_pane_height: u16,
-    pub explorer_content_width: usize,
-    pub inspector_pane_height: u16,
-    pub result_pane_height: u16,
-    pub command_line_visible_width: Option<usize>,
-    pub connection_list_pane_height: Option<u16>,
-    pub table_picker_pane_height: Option<u16>,
-    pub table_picker_filter_visible_width: Option<usize>,
-    pub er_picker_pane_height: Option<u16>,
-    pub er_picker_filter_visible_width: Option<usize>,
-    pub query_history_picker_pane_height: Option<u16>,
-    pub query_history_picker_filter_visible_width: Option<usize>,
-    pub jsonb_detail_editor_visible_rows: Option<usize>,
-    pub confirm_preview_viewport_height: Option<u16>,
-    pub confirm_preview_content_height: Option<u16>,
-    pub confirm_preview_scroll: u16,
-    pub explain_compare_viewport_height: Option<u16>,
-}
 
 pub trait Renderer {
     fn draw(
